refactor(container): mv Article-xxx to unit scope
