+++ src/containers/content/VideoContent/SideCards.js
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import Informer from '@/containers/tool/Informer'
-import ArticleAuthorCard from '@/containers/ArticleAuthorCard'
+import ArticleAuthorCard from '@/containers/unit/ArticleAuthorCard'
 import ContentSourceCard from '@/components/ContentSourceCard'
 
 import { Wrapper, ReportWrapper } from './styles/side_cards'
