+++ src/containers/content/VideoContent/index.js
@@ -9,9 +9,9 @@ import React from 'react'
 import { THREAD } from '@/constant'
 import { connectStore, buildLog } from '@/utils'
 
-import ArticleBodyHeader from '@/containers/ArticleBodyHeader'
+import ArticleBodyHeader from '@/containers/unit/ArticleBodyHeader'
 import Comments from '@/containers/unit/Comments'
-import ArticleAuthorCard from '@/containers/ArticleAuthorCard'
+import ArticleAuthorCard from '@/containers/unit/ArticleAuthorCard'
 import ContentSourceCard from '@/components/ContentSourceCard'
 
 import Sticky from '@/components/Sticky'
