+++ src/containers/viewer/JobViewer/index.js
@@ -10,8 +10,8 @@ import { THREAD } from '@/constant'
 import { connectStore, buildLog } from '@/utils'
 
 import Comments from '@/containers/unit/Comments'
-import ArticleViewerHeader from '@/containers/ArticleViewerHeader'
-import ArticleBodyHeader from '@/containers/ArticleBodyHeader'
+import ArticleViewerHeader from '@/containers/unit/ArticleViewerHeader'
+import ArticleBodyHeader from '@/containers/unit/ArticleBodyHeader'
 import Maybe from '@/components/Maybe'
 import MarkDownRender from '@/components/MarkDownRender'
 import { ArticleContentLoading } from '@/components/LoadingEffects'
