+++ src/containers/viewer/VideoViewer/index.js
@@ -9,8 +9,8 @@ import React from 'react'
 import { THREAD } from '@/constant'
 import { connectStore, buildLog } from '@/utils'
 
-import ArticleViewerHeader from '@/containers/ArticleViewerHeader'
-import ArticleBodyHeader from '@/containers/ArticleBodyHeader'
+import ArticleViewerHeader from '@/containers/unit/ArticleViewerHeader'
+import ArticleBodyHeader from '@/containers/unit/ArticleBodyHeader'
 import Comments from '@/containers/unit/Comments'
 import Maybe from '@/components/Maybe'
 import VideoPoster from '@/components/VideoPoster'
